[get sources] Update sources

+++ codeready-workspaces-devworkspace-controller/get-sources-jenkins.sh
@@ -5,6 +5,7 @@ JOB_BRANCH=""
 doRhpkgContainerBuild=1
 forceBuild=0
 forcePull=0
+doPull=1
 verbose=0
 
 # specific flags for sync.sh
@@ -16,6 +17,7 @@ while [[ "$#" -gt 0 ]]; do
     '-n'|'--nobuild') doRhpkgContainerBuild=0; shift 0;;
     '-f'|'--force-build') forceBuild=1; shift 0;;
     '-p'|'--force-pull') forcePull=1; shift 0;;
+    '-np'|'--no-pull') doPull=0; shift 0;;
     '-s'|'--scratch') scratchFlag="--scratch"; shift 0;;
     '--source-branch'|'-sb') SOURCE_BRANCH="$2"; shift 1;;
     '--no-vendor'|'-nv') UPDATE_VENDOR="--no-vendor";;
@@ -31,29 +33,30 @@ function log()
   fi
 }
 
-if [[ ! ${JOB_BRANCH} ]]; then
-  MIDSTM_BRANCH=$(git rev-parse --abbrev-ref HEAD)
-else
-  MIDSTM_BRANCH="crw-${JOB_BRANCH}-rhel-8"; if [[ ${JOB_BRANCH} == "2.x" ]]; then MIDSTM_BRANCH="crw-2-rhel-8"; fi
-CSV_VERSION=$(curl -sSLo- https://raw.githubusercontent.com/redhat-developer/codeready-workspaces/${MIDSTM_BRANCH}/dependencies/VERSION)
+if [[ ${doPull} -eq 1 ]]; then
+  if [[ ! ${JOB_BRANCH} ]]; then
+    MIDSTM_BRANCH=$(git rev-parse --abbrev-ref HEAD)
+  else
+    MIDSTM_BRANCH="crw-${JOB_BRANCH}-rhel-8"; if [[ ${JOB_BRANCH} == "2.x" ]]; then MIDSTM_BRANCH="crw-2-rhel-8"; fi
+  CSV_VERSION=$(curl -sSLo- https://raw.githubusercontent.com/redhat-developer/codeready-workspaces/${MIDSTM_BRANCH}/dependencies/VERSION)
 
+  fi
+  CSV_VERSION=$(curl -sSLo- https://raw.githubusercontent.com/redhat-developer/codeready-workspaces/${MIDSTM_BRANCH}/dependencies/VERSION)
+
+  SCRIPTS_DIR=$(cd "$(dirname "$0")"; pwd)
+
+  # fetch upstream sources into temp folder
+  TMPDIR=$(mktemp -d)
+  pushd $TMPDIR >/dev/null
+    git clone https://github.com/devfile/devworkspace-operator.git tmp && cd tmp
+    git checkout ${SOURCE_BRANCH}
+  popd >/dev/null
+
+  chmod +x ./build/scripts/sync.sh
+  ./build/scripts/sync.sh -v ${CSV_VERSION}.0 -s ${TMPDIR}/tmp -t ${SCRIPTS_DIR} ${UPDATE_VENDOR}
+  # cleanup
+  rm -fr ${TMPDIR} bin/ bootstrap.Dockerfile 
 fi
-CSV_VERSION=$(curl -sSLo- https://raw.githubusercontent.com/redhat-developer/codeready-workspaces/${MIDSTM_BRANCH}/dependencies/VERSION)
-
-SCRIPTS_DIR=$(cd "$(dirname "$0")"; pwd)
-
-# fetch upstream sources into temp folder
-TMPDIR=$(mktemp -d)
-pushd $TMPDIR >/dev/null
-  git clone https://github.com/devfile/devworkspace-operator.git tmp && cd tmp
-  git checkout ${SOURCE_BRANCH}
-popd >/dev/null
-
-chmod +x ./build/scripts/sync.sh
-./build/scripts/sync.sh -v ${CSV_VERSION}.0 -s ${TMPDIR}/tmp -t ${SCRIPTS_DIR} ${UPDATE_VENDOR}
-
-# cleanup
-rm -fr ${TMPDIR} bin/ bootstrap.Dockerfile 
 
 git update-index --refresh || true  # ignore timestamp updates
 if [[ $(git diff-index HEAD --) ]] || [[ ${forcePull} -ne 0 ]]; then # file changed
